+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Updates option name: `site-max` and `siteMax` are now `max` (#79)
 - Initial support for outputting a single value using a `min()` function, rather than a media query (#79)
 - The `max` option can now be set to a `var()` function (#79)
+- Properties are deprecated and will be removed in a future version: `tidy-column`, `tidy-offset`, `tidy-span`, `tidy-offset-left`, `tidy-offset-right` (#81)
 
 **Removed**